refactor: ti: segregate scmi clk and pd data into soc folder
Seperate out the pd and clk arrays for AM62L into it's own folder so that the
scmi drivers can be generic enough and pick the soc specific arrays from the
soc folder data files.
