+++ plat/ti/k3/board/am62l/board.mk
@@ -21,6 +21,9 @@ include ${PLAT_PATH}/common/drivers/scmi/ti_scmi.mk
 include ${PLAT_PATH}/common/drivers/pm/pm.mk
 include ${PLAT_PATH}/board/am62l/pm/soc_pm.mk
 
+PLAT_INCLUDES += -Iplat/ti/k3/board/${TARGET_BOARD}/pm			\
+		 -I${PLAT_PATH}/board/am62l/scmi			\
+
 BL31_SOURCES		+=	\
 				${PLAT_PATH}/common/k3_svc.c		\
 				drivers/scmi-msg/base.c			\
