+++ plat/ti/k3/common/drivers/scmi/scmi_clock.c
@@ -18,526 +18,7 @@
 #include <clocks.h>
 #include <device_clk.h>
 
-
-#define AM62L_SCMI_CLOCK(_dev_id, _clock_id, _name)		\
-	{							\
-		.dev_id	= _dev_id,				\
-		.name = appen(xstr(_clock_id), "_", _name ),	\
-		.clock_id = _clock_id,				\
-		.rates = am62l_rates,				\
-	}
-
-#define xstr(s) str(s)
-#define str(s) #s
-
-#define appen(a,b,c) a b c
-
-#define AM62L_SCMI_CLOCK_MUX AM62L_SCMI_CLOCK
-#define AM62L_SCMI_CLOCK_PARENT AM62L_SCMI_CLOCK
-#define AM62L_SCMI_CLOCK_OUTPUT AM62L_SCMI_CLOCK
-#define AM62L_SCMI_CLOCK_MUX AM62L_SCMI_CLOCK
-
-/*
- * Our HSDIVs can support any range of divisions, so it's difficult
- * to derive a fix number of clock rates array per device.
- * There isn't as such a fixed rates array per device.
- * Hence, use a range from 0 to maximum supported clock frequency.
- */
-static unsigned long am62l_rates[] = {
-	0, 2000000000, 1,
-};
-
-ti_scmi_clock_t clock_table[] = {
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_ADC_CLK, str(ADC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_ADC_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(ADC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_ADC_CLK_PARENT_SAM62_PLL_CTRL_WRAP_WKUP_0_CHIP_DIV1_CLK_CLK12,str(ADC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_ADC_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT8_CLK,str(ADC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_ADC_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(ADC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_SYS_CLK,str(ADC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ADC0, AM62LX_DEV_ADC0_VBUS_CLK,str(ADC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MAIN_GPIOMUX_INTROUTER0, AM62LX_DEV_MAIN_GPIOMUX_INTROUTER0_INTR_CLK,str(MAIN_GPIOMUX_INTROUTER0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TIMESYNC_INTROUTER0, AM62LX_DEV_TIMESYNC_INTROUTER0_INTR_CLK,str(TIMESYNC_INTROUTER0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPPI_CLK_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_BOARD_0_CP_GEMAC_CPTS0_RFT_CLK_OUT,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_SAM62_PLL_CTRL_WRAP_WKUP_0_CHIP_DIV1_CLK_CLK, str(CPSW0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_RFT_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT0_CLK, str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_GMII1_MR_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_GMII1_MT_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_GMII2_MR_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_GMII2_MT_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_GMII_RFT_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_RGMII_MHZ_250_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_RGMII_MHZ_50_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_RGMII_MHZ_5_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_RMII1_MHZ_50_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_RMII2_MHZ_50_CLK,str(CPSW0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_GENF0,str(CPSW0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_CPTS_GENF1,str(CPSW0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_CPSW0, AM62LX_DEV_CPSW0_MDIO_MDCLK_O,str(CPSW0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPT2_AGGR0, AM62LX_DEV_CPT2_AGGR0_VCLK_CLK,str(CPT2_AGGR0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_CPT2_AGGR1, AM62LX_DEV_CPT2_AGGR1_VCLK_CLK,str(CPT2_AGGR1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_CPT2_AGGR0, AM62LX_DEV_WKUP_CPT2_AGGR0_VCLK_CLK,str(WKUP_CPT2_AGGR0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_STM0, AM62LX_DEV_STM0_ATB_CLK,str(STM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_STM0, AM62LX_DEV_STM0_CORE_CLK,str(STM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_STM0, AM62LX_DEV_STM0_VBUSP_CLK,str(STM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_ATB_CLK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_CORE_CLK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_JTAG_TCK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_P1500_WRCK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_TREXPT_CLK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_DEBUGSS_WRAP0, AM62LX_DEV_DEBUGSS_WRAP0_CSTPIU_TRACECLK,str(DEBUGSS_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DMASS0_BCDMA_0, AM62LX_DEV_DMASS0_BCDMA_0_CLK,str(DMASS0_BCDMA_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DMASS0_PKTDMA_0, AM62LX_DEV_DMASS0_PKTDMA_0_CLK,str(DMASS0_PKTDMA_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_HCLK_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF0,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF1,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_BOARD_0_CP_GEMAC_CPTS0_RFT_CLK_OUT,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT1_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_TCLK_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(TIMER0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_TIMER0, AM62LX_DEV_TIMER0_TIMER_PWM,str(TIMER0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TIMER1, AM62LX_DEV_TIMER1_TIMER_HCLK_CLK,str(TIMER1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_TIMER1, AM62LX_DEV_TIMER1_TIMER_TCLK_CLK,str(TIMER1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER1, AM62LX_DEV_TIMER1_TIMER_TCLK_CLK_PARENT_MAIN_TIMERCLKN_SEL_OUT1,str(TIMER1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER1, AM62LX_DEV_TIMER1_TIMER_TCLK_CLK_PARENT_DMTIMER_DMC1MS_MAIN_0_TIMER_PWM,str(TIMER1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_TIMER1, AM62LX_DEV_TIMER1_TIMER_PWM,str(TIMER1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_HCLK_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF0,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF1,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_BOARD_0_CP_GEMAC_CPTS0_RFT_CLK_OUT,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT1_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_TCLK_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(TIMER2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_TIMER2, AM62LX_DEV_TIMER2_TIMER_PWM,str(TIMER2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TIMER3, AM62LX_DEV_TIMER3_TIMER_HCLK_CLK,str(TIMER3)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_TIMER3, AM62LX_DEV_TIMER3_TIMER_TCLK_CLK,str(TIMER3)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER3, AM62LX_DEV_TIMER3_TIMER_TCLK_CLK_PARENT_MAIN_TIMERCLKN_SEL_OUT3,str(TIMER3)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_TIMER3, AM62LX_DEV_TIMER3_TIMER_TCLK_CLK_PARENT_DMTIMER_DMC1MS_MAIN_2_TIMER_PWM,str(TIMER3)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_TIMER3, AM62LX_DEV_TIMER3_TIMER_PWM,str(TIMER3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_HCLK_CLK,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_SAM62_PLL_CTRL_WRAP_WKUP_0_CHIP_DIV1_CLK_CLK2,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF0,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_TCLK_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_WKUP_TIMER0, AM62LX_DEV_WKUP_TIMER0_TIMER_PWM,str(WKUP_TIMER0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_TIMER1, AM62LX_DEV_WKUP_TIMER1_TIMER_HCLK_CLK,str(WKUP_TIMER1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_TIMER1, AM62LX_DEV_WKUP_TIMER1_TIMER_TCLK_CLK,str(WKUP_TIMER1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER1, AM62LX_DEV_WKUP_TIMER1_TIMER_TCLK_CLK_PARENT_WKUP_TIMERCLKN_SEL_OUT1,str(WKUP_TIMER1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_TIMER1, AM62LX_DEV_WKUP_TIMER1_TIMER_TCLK_CLK_PARENT_DMTIMER_DMC1MS_WKUP_0_TIMER_PWM,str(WKUP_TIMER1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ECAP0, AM62LX_DEV_ECAP0_VBUS_CLK,str(ECAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ECAP1, AM62LX_DEV_ECAP1_VBUS_CLK,str(ECAP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ECAP2, AM62LX_DEV_ECAP2_VBUS_CLK,str(ECAP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ELM0, AM62LX_DEV_ELM0_VBUSP_CLK,str(ELM0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_IO_CLK_I,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC1_CLKLB_OUT,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC1_CLK_OUT,str(MMCSD1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_VBUS_CLK,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_XIN_CLK,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT5_CLK,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT9_CLK,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MMCSD1, AM62LX_DEV_MMCSD1_EMMCSDSS_IO_CLK_O,str(MMCSD1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_IO_CLK_I,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC2_CLKLB_OUT,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC2_CLK_OUT,str(MMCSD2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_VBUS_CLK,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_XIN_CLK,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT5_CLK,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT9_CLK,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MMCSD2, AM62LX_DEV_MMCSD2_EMMCSDSS_IO_CLK_O,str(MMCSD2)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_IO_CLK_I,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC0_CLKLB_OUT,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_IO_CLK_I_PARENT_BOARD_0_MMC0_CLK_OUT,str(MMCSD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_VBUS_CLK,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_XIN_CLK,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT5_CLK,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_XIN_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT9_CLK,str(MMCSD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MMCSD0, AM62LX_DEV_MMCSD0_EMMCSDSS_IO_CLK_O,str(MMCSD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EQEP0, AM62LX_DEV_EQEP0_VBUS_CLK,str(EQEP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EQEP1, AM62LX_DEV_EQEP1_VBUS_CLK,str(EQEP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EQEP2, AM62LX_DEV_EQEP2_VBUS_CLK,str(EQEP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_DQS_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_ICLK_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_ICLK_CLK_PARENT_BOARD_0_OSPI0_DQS_OUT,str(FSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_ICLK_CLK_PARENT_BOARD_0_OSPI0_LBCLKO_OUT,str(FSS0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_RCLK_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_RCLK_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT7_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_RCLK_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT4_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_VBUS_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_FSS0, AM62LX_DEV_FSS0_OSPI0_OCLK_CLK,str(FSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_GICSS0, AM62LX_DEV_GICSS0_VCLK_CLK,str(GICSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_GPIO0, AM62LX_DEV_GPIO0_MMR_CLK,str(GPIO0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_GPIO2, AM62LX_DEV_GPIO2_MMR_CLK,str(GPIO2)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_GPIO0, AM62LX_DEV_WKUP_GPIO0_MMR_CLK,str(WKUP_GPIO0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GPIO0, AM62LX_DEV_WKUP_GPIO0_MMR_CLK_PARENT_SAM62_PLL_CTRL_WRAP_WKUP_0_CHIP_DIV1_CLK_CLK4,str(WKUP_GPIO0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GPIO0, AM62LX_DEV_WKUP_GPIO0_MMR_CLK_PARENT_RTCSS_WKUP_0_OSC_32K_CLK,str(WKUP_GPIO0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GPIO0, AM62LX_DEV_WKUP_GPIO0_MMR_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(WKUP_GPIO0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GPIO0, AM62LX_DEV_WKUP_GPIO0_MMR_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(WKUP_GPIO0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_FUNC_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_FUNC_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT3_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_FUNC_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT3_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_PI_GPMC_RET_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_VBUSM_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_GPMC0, AM62LX_DEV_GPMC0_PO_GPMC_DEV_CLK,str(GPMC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_DPHY_0_RX_ESC_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_DPHY_0_TX_ESC_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_DPI_0_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_PLL_CTRL_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_PPI_0_TXBYTECLKHS_CL_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS_DSI0, AM62LX_DEV_DSS_DSI0_SYS_CLK,str(DSS_DSI0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_DSS0, AM62LX_DEV_DSS0_DPI_0_IN_CLK,str(DSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_DSS0, AM62LX_DEV_DSS0_DPI_0_IN_CLK_PARENT_HSDIV0_16FFT_MAIN_17_HSDIVOUT0_CLK,str(DSS0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_DSS0, AM62LX_DEV_DSS0_DPI_0_IN_CLK_PARENT_BOARD_0_VOUT0_EXTPCLKIN_OUT,str(DSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DSS0, AM62LX_DEV_DSS0_DSS_FUNC_CLK,str(DSS0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_DSS0, AM62LX_DEV_DSS0_DPI_0_OUT_CLK,str(DSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EPWM0, AM62LX_DEV_EPWM0_VBUSP_CLK,str(EPWM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EPWM1, AM62LX_DEV_EPWM1_VBUSP_CLK,str(EPWM1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_EPWM2, AM62LX_DEV_EPWM2_VBUSP_CLK,str(EPWM2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_LED0, AM62LX_DEV_LED0_VBUS_CLK,str(LED0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_PBIST0, AM62LX_DEV_PBIST0_CLK8_CLK,str(PBIST0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_PBIST0, AM62LX_DEV_PBIST0_TCLK_CLK,str(PBIST0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_PBIST0, AM62LX_DEV_WKUP_PBIST0_CLK8_CLK,str(WKUP_PBIST0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_VTM0, AM62LX_DEV_WKUP_VTM0_FIX_REF2_CLK,str(WKUP_VTM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_VTM0, AM62LX_DEV_WKUP_VTM0_FIX_REF_CLK,str(WKUP_VTM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_VTM0, AM62LX_DEV_WKUP_VTM0_VBUSP_CLK,str(WKUP_VTM0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_CCLK_CLK,str(MCAN0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_CCLK_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT4_CLK,str(MCAN0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_CCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(MCAN0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_CCLK_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(MCAN0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_CCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(MCAN0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCAN0, AM62LX_DEV_MCAN0_MCANSS_HCLK_CLK,str(MCAN0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_CCLK_CLK,str(MCAN1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_CCLK_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT4_CLK,str(MCAN1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_CCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(MCAN1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_CCLK_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(MCAN1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_CCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(MCAN1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCAN1, AM62LX_DEV_MCAN1_MCANSS_HCLK_CLK,str(MCAN1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_CCLK_CLK,str(MCAN2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_CCLK_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT4_CLK,str(MCAN2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_CCLK_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(MCAN2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_CCLK_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(MCAN2)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_CCLK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(MCAN2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCAN2, AM62LX_DEV_MCAN2_MCANSS_HCLK_CLK,str(MCAN2)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_AUX_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_AUX_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT8_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_AUX_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT1_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_ACLKR_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_ACLKX_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AFSR_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AFSX_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_PIN_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_PIN_PARENT_GLUELOGIC_HFOSC0_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_PIN_PARENT_BOARD_0_AUDIO_EXT_REFCLK0_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_PIN_PARENT_BOARD_0_AUDIO_EXT_REFCLK1_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_PIN,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_PIN_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_PIN_PARENT_GLUELOGIC_HFOSC0_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_PIN_PARENT_BOARD_0_AUDIO_EXT_REFCLK0_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_PIN_PARENT_BOARD_0_AUDIO_EXT_REFCLK1_OUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_VBUSP_CLK,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_ACLKR_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_ACLKX_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AFSR_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AFSX_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKR_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP0, AM62LX_DEV_MCASP0_MCASP_AHCLKX_POUT,str(MCASP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_AUX_CLK,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_ACLKR_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_ACLKX_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AFSR_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AFSX_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AHCLKR_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AHCLKX_PIN,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_VBUSP_CLK,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_ACLKR_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_ACLKX_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AFSR_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AFSX_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AHCLKR_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP1, AM62LX_DEV_MCASP1_MCASP_AHCLKX_POUT,str(MCASP1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_AUX_CLK,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_ACLKR_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_ACLKX_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AFSR_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AFSX_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AHCLKR_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AHCLKX_PIN,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_VBUSP_CLK,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_ACLKR_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_ACLKX_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AFSR_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AFSX_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AHCLKR_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCASP2, AM62LX_DEV_MCASP2_MCASP_AHCLKX_POUT,str(MCASP2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C0, AM62LX_DEV_I2C0_CLK,str(I2C0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C0, AM62LX_DEV_I2C0_PISCL,str(I2C0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C0, AM62LX_DEV_I2C0_PISYS_CLK,str(I2C0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_I2C0, AM62LX_DEV_I2C0_PORSCL,str(I2C0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C1, AM62LX_DEV_I2C1_CLK,str(I2C1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C1, AM62LX_DEV_I2C1_PISCL,str(I2C1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C1, AM62LX_DEV_I2C1_PISYS_CLK,str(I2C1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_I2C1, AM62LX_DEV_I2C1_PORSCL,str(I2C1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C2, AM62LX_DEV_I2C2_CLK,str(I2C2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C2, AM62LX_DEV_I2C2_PISCL,str(I2C2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C2, AM62LX_DEV_I2C2_PISYS_CLK,str(I2C2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_I2C2, AM62LX_DEV_I2C2_PORSCL,str(I2C2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C3, AM62LX_DEV_I2C3_CLK,str(I2C3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C3, AM62LX_DEV_I2C3_PISCL,str(I2C3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_I2C3, AM62LX_DEV_I2C3_PISYS_CLK,str(I2C3)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_I2C3, AM62LX_DEV_I2C3_PORSCL,str(I2C3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_I2C0, AM62LX_DEV_WKUP_I2C0_CLK,str(WKUP_I2C0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_I2C0, AM62LX_DEV_WKUP_I2C0_PISCL,str(WKUP_I2C0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_I2C0, AM62LX_DEV_WKUP_I2C0_PISYS_CLK,str(WKUP_I2C0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_WKUP_I2C0, AM62LX_DEV_WKUP_I2C0_PORSCL,str(WKUP_I2C0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_GTC0, AM62LX_DEV_WKUP_GTC0_GTC_CLK,str(WKUP_GTC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTC0, AM62LX_DEV_WKUP_GTC0_GTC_CLK_PARENT_WKUP_GTCCLK_SEL_OUT0,str(WKUP_GTC0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTC0, AM62LX_DEV_WKUP_GTC0_GTC_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(WKUP_GTC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_GTC0, AM62LX_DEV_WKUP_GTC0_VBUSP_CLK,str(WKUP_GTC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_RTCSS0, AM62LX_DEV_WKUP_RTCSS0_ANA_OSC32K_CLK,str(WKUP_RTCSS0)), /* lfosc is not suppported o vlab */
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_RTCSS0, AM62LX_DEV_WKUP_RTCSS0_AUX_32K_CLK,str(WKUP_RTCSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_RTCSS0, AM62LX_DEV_WKUP_RTCSS0_JTAG_WRCK,str(WKUP_RTCSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_RTCSS0, AM62LX_DEV_WKUP_RTCSS0_VCLK_CLK,str(WKUP_RTCSS0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_WKUP_RTCSS0, AM62LX_DEV_WKUP_RTCSS0_OSC_32K_CLK,str(WKUP_RTCSS0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_RTI_CLK,str(RTI0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_RTI_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(RTI0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_RTI_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(RTI0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_RTI_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(RTI0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_RTI_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(RTI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_RTI0, AM62LX_DEV_RTI0_VBUSP_CLK,str(RTI0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_RTI_CLK,str(RTI1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_RTI_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(RTI1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_RTI_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(RTI1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_RTI_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(RTI1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_RTI_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(RTI1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_RTI1, AM62LX_DEV_RTI1_VBUSP_CLK,str(RTI1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS0, AM62LX_DEV_DEBUGSS0_CFG_CLK,str(DEBUGSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS0, AM62LX_DEV_DEBUGSS0_DBG_CLK,str(DEBUGSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DEBUGSS0, AM62LX_DEV_DEBUGSS0_SYS_CLK,str(DEBUGSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MSRAM_96K0, AM62LX_DEV_MSRAM_96K0_VCLK_CLK,str(MSRAM_96K0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_PSRAM_64K0, AM62LX_DEV_WKUP_PSRAM_64K0_CLK_CLK,str(WKUP_PSRAM_64K0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_ROM0, AM62LX_DEV_ROM0_CLK_CLK,str(ROM0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_PSC0, AM62LX_DEV_PSC0_CLK,str(PSC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_PSC0, AM62LX_DEV_PSC0_SLOW_CLK,str(PSC0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_DEEPSLEEP_SOURCES0, AM62LX_DEV_WKUP_DEEPSLEEP_SOURCES0_CLK_12M_RC_CLK,str(WKUP_DEEPSLEEP_SOURCES0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0, AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0_COREPAC_ARM_CLK_CLK,str(ARM_COREPACK_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0, AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0_PLL_CTRL_CLK ,str(ARM_COREPACK_0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0, AM62LX_DEV_COMPUTE_CLUSTER0_ARM_COREPACK_0_A53_DIVH_CLK4_OBSCLK_OUT_CLK,str(ARM_COREPACK_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_DFTSS0, AM62LX_DEV_WKUP_DFTSS0_PLL_CLK,str(WKUP_DFTSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_DFTSS0, AM62LX_DEV_WKUP_DFTSS0_VBUSP_CLK_CLK,str(WKUP_DFTSS0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI0, AM62LX_DEV_MCSPI0_CLKSPIREF_CLK,str(MCSPI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI0, AM62LX_DEV_MCSPI0_VBUSP_CLK,str(MCSPI0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCSPI0, AM62LX_DEV_MCSPI0_IO_CLKSPIO_CLK,str(MCSPI0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI1, AM62LX_DEV_MCSPI1_CLKSPIREF_CLK,str(MCSPI1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI1, AM62LX_DEV_MCSPI1_VBUSP_CLK,str(MCSPI1)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCSPI1, AM62LX_DEV_MCSPI1_IO_CLKSPIO_CLK,str(MCSPI1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI2, AM62LX_DEV_MCSPI2_CLKSPIREF_CLK,str(MCSPI2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI2, AM62LX_DEV_MCSPI2_VBUSP_CLK,str(MCSPI2)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCSPI2, AM62LX_DEV_MCSPI2_IO_CLKSPIO_CLK,str(MCSPI2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI3, AM62LX_DEV_MCSPI3_CLKSPIREF_CLK,str(MCSPI3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_MCSPI3, AM62LX_DEV_MCSPI3_VBUSP_CLK,str(MCSPI3)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_MCSPI3, AM62LX_DEV_MCSPI3_IO_CLKSPIO_CLK,str(MCSPI3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_TRNG_DRBG_EIP76D_WRAP0, AM62LX_DEV_TRNG_DRBG_EIP76D_WRAP0_VCLK_CLK,str(TRNG_DRBG_EIP76D_WRAP0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART1, AM62LX_DEV_UART1_FCLK_CLK,str(UART1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART1, AM62LX_DEV_UART1_VBUSP_CLK,str(UART1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART2, AM62LX_DEV_UART2_FCLK_CLK,str(UART2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART2, AM62LX_DEV_UART2_VBUSP_CLK,str(UART2)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART3, AM62LX_DEV_UART3_FCLK_CLK,str(UART3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART3, AM62LX_DEV_UART3_VBUSP_CLK,str(UART3)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART4, AM62LX_DEV_UART4_FCLK_CLK,str(UART4)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART4, AM62LX_DEV_UART4_VBUSP_CLK,str(UART4)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART5, AM62LX_DEV_UART5_FCLK_CLK,str(UART5)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART5, AM62LX_DEV_UART5_VBUSP_CLK,str(UART5)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART6, AM62LX_DEV_UART6_FCLK_CLK,str(UART6)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART6, AM62LX_DEV_UART6_VBUSP_CLK,str(UART6)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_UART0, AM62LX_DEV_WKUP_UART0_FCLK_CLK,str(WKUP_UART0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_WKUP_UART0, AM62LX_DEV_WKUP_UART0_VBUSP_CLK,str(WKUP_UART0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB0, AM62LX_DEV_USB0_BUS_CLK,str(USB0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB0, AM62LX_DEV_USB0_CFG_CLK,str(USB0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB0, AM62LX_DEV_USB0_USB2_APB_PCLK_CLK,str(USB0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_USB0, AM62LX_DEV_USB0_USB2_REFCLOCK_CLK,str(USB0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_USB0, AM62LX_DEV_USB0_USB2_REFCLOCK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(USB0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_USB0, AM62LX_DEV_USB0_USB2_REFCLOCK_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT8_CLK4,str(USB0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB0, AM62LX_DEV_USB0_USB2_TAP_TCK,str(USB0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB1, AM62LX_DEV_USB1_BUS_CLK,str(USB1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB1, AM62LX_DEV_USB1_CFG_CLK,str(USB1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB1, AM62LX_DEV_USB1_USB2_APB_PCLK_CLK,str(USB1)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_USB1, AM62LX_DEV_USB1_USB2_REFCLOCK_CLK,str(USB1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_USB1, AM62LX_DEV_USB1_USB2_REFCLOCK_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(USB1)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_USB1, AM62LX_DEV_USB1_USB2_REFCLOCK_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT8_CLK4,str(USB1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_USB1, AM62LX_DEV_USB1_USB2_TAP_TCK,str(USB1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_DPHY_REF_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_DPHY_REF_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_DPHY_REF_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT8_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP1_PPI_M_TXCLKESC_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP2_PPI_M_TXCLKESC_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP3_PPI_M_TXCLKESC_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP4_PPI_M_TXCLKESC_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_PSM_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP1_PPI_M_RXCLKESC_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_DPHY_TX0, AM62LX_DEV_DPHY_TX0_IP1_PPI_TXBYTECLKHS_CL_CLK,str(DPHY_TX0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0, AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0_FUNC_CLKIN_CLK,str(CLKDIV_0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0, AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0_DIVH_CLK4_CLK_CLK,str(CLKDIV_0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0, AM62LX_DEV_COMPUTE_CLUSTER0_CLKDIV_0_DIVP_CLK1_CLK_CLK,str(CLKDIV_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_PBIST_0, AM62LX_DEV_COMPUTE_CLUSTER0_PBIST_0_DIVH_CLK4_CLK_CLK,str(PBIST_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_PBIST_0, AM62LX_DEV_COMPUTE_CLUSTER0_PBIST_0_DIVP_CLK1_CLK_CLK,str(PBIST_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_A53_0, AM62LX_DEV_COMPUTE_CLUSTER0_A53_0_A53_CORE0_ARM_CLK_CLK,str(A53_0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_COMPUTE_CLUSTER0_A53_1, AM62LX_DEV_COMPUTE_CLUSTER0_A53_1_A53_CORE1_ARM_CLK_CLK,str(A53_1)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART0, AM62LX_DEV_UART0_FCLK_CLK,str(UART0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_UART0, AM62LX_DEV_UART0_VBUSP_CLK,str(UART0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_AUDIO_EXT_REFCLK0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_AUDIO_EXT_REFCLK1_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_CLKOUT0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_CLKOUT0_IN_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK5,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_CLKOUT0_IN_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK10,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_GPMC0_CLKLB_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_GPMC0_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_GPMC0_FCLK_MUX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C0_SCL_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C1_SCL_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C2_SCL_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C3_SCL_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_ACLKR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_ACLKX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_AFSR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_AFSX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_ACLKR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_ACLKX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_AFSR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_AFSX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_ACLKR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_ACLKX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_AFSR_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_AFSX_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MDIO0_MDC_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC0_CLKLB_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC0_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC1_CLKLB_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC1_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC2_CLKLB_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC2_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OBSCLK0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OBSCLK1_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OSPI0_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OSPI0_LBCLKO_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_SPI0_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_SPI1_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_SPI2_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_SPI3_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TIMER_IO0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TIMER_IO1_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TIMER_IO2_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TIMER_IO3_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TRC_CLK_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_CLKOUT0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_CLKOUT0_IN_PARENT_WKUP_CLKOUT_SEL_OUT0,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_CLKOUT0_IN_PARENT_GLUELOGIC_HFOSC0_CLK,str(BOARD0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_OBSCLK0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_OBSCLK0_IN_PARENT_WKUP_OBSCLK_MUX_SEL_OUT0,str(BOARD0)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_OBSCLK0_IN_PARENT_GLUELOGIC_HFOSC0_CLK,str(BOARD0)),
-	AM62L_SCMI_CLOCK(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_SYSCLKOUT0_IN,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_AUDIO_EXT_REFCLK0_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_AUDIO_EXT_REFCLK1_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_CP_GEMAC_CPTS0_RFT_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_EXT_REFCLK1_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_GPMC0_CLKLB_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C0_SCL_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C1_SCL_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C2_SCL_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_I2C3_SCL_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_ACLKR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_ACLKX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_AFSR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP0_AFSX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_ACLKR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_ACLKX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_AFSR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP1_AFSX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_ACLKR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_ACLKX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_AFSR_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MCASP2_AFSX_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC0_CLKLB_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC0_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC1_CLKLB_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC1_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC2_CLKLB_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_MMC2_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OSPI0_DQS_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_OSPI0_LBCLKO_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_RMII1_REF_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_RMII2_REF_CLK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_TCK_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_VOUT0_EXTPCLKIN_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_EXT_REFCLK0_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_OUTPUT(AM62LX_DEV_BOARD0, AM62LX_DEV_BOARD0_WKUP_I2C0_SCL_OUT,str(BOARD0)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_POSTDIV4_16FF_MAIN_0_HSDIVOUT6_CLK,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_BOARD_0_CP_GEMAC_CPTS0_RFT_CLK_OUT,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_BOARD_0_WKUP_EXT_REFCLK0_OUT,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_BOARD_0_EXT_REFCLK1_OUT,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD, AM62LX_DEV_WKUP_GTCCLK_SEL_DEV_VD_CLK_PARENT_SAM62_PLL_CTRL_WRAP_WKUP_0_CHIP_DIV1_CLK_CLK2,str(WKUP_GTCCLK_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT8_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT0_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT4_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT0_CLK_DUP0,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_WKUP_32KHZ_GEN_0_HSDIVOUT0_CLK8,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_WKUP_0_HSDIVOUT0_CLK,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD, AM62LX_DEV_WKUP_OBSCLK_MUX_SEL_DEV_VD_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(WKUP_OBSCLK_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_RTCSS_WKUP_0_OSC_32K_CLK,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT7_CLK,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_WKUP_0_HSDIVOUT0_CLK2,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_POSTDIV4_16FF_WKUP_0_HSDIVOUT8_CLK,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_CLK_32K_RC_SEL_OUT0,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD, AM62LX_DEV_WKUP_CLKOUT_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(WKUP_CLKOUT_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT0_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF1,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_MAIN_17_HSDIVOUT0_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_SAM62L_A53_256KB_WRAP_MAIN_0_ARM_COREPACK_0_A53_DIVH_CLK4_OBSCLK_OUT_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_MAIN_0_HSDIVOUT2_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLKOUT,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_WKUP_32KHZ_GEN_0_HSDIVOUT0_CLK8,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_HSDIV4_16FFT_WKUP_0_HSDIVOUT0_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_HFOSC0_CLK,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD, AM62LX_DEV_OBSCLK0_MUX_SEL_DEV_VD_CLK_PARENT_CPSW_3GUSS_AM62L_MAIN_0_CPTS_GENF0,str(OBSCLK0_MUX_SEL_DEV_VD)),
-	AM62L_SCMI_CLOCK_MUX(AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD, AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD_CLK, str(CLK_32K_RC)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD, AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3, str(CLK_32K_RC)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD, AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD_CLK_PARENT_HSDIV0_16FFT_WKUP_32KHZ_GEN_0_HSDIVOUT0_CLK8, str(CLK_32K_RC)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD, AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD_CLK_PARENT_GLUELOGIC_RCOSC_CLK_1P0V_97P65K3_DUP0, str(CLK_32K_RC)),
-	AM62L_SCMI_CLOCK_PARENT(AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD, AM62LX_DEV_CLK_32K_RC_SEL_DEV_VD_CLK_PARENT_RTCSS_WKUP_0_OSC_32K_CLK, str(CLK_32K_RC)),
-};
+#include <scmi_clk_data.h>
 
 ti_scmi_clock_t *ti_scmi_get_clock(uint32_t agent_id __unused,
 				   uint32_t clock_id)
